benchmark refactor
